implementation of enums in regex, distribution of validations. low coupling.
