+++ src/app/core/util/validators/password.validators.ts
@@ -1,21 +1,4 @@
-
 import { AbstractControl, ValidationErrors, ValidatorFn } from "@angular/forms";
-
-
-export function patternValidator(regex: RegExp, error: ValidationErrors): ValidatorFn {
-    return (control: AbstractControl): ValidationErrors | null => {
-        if (!control.value) {
-        // if control is empty return no error
-            return null;
-        }
-    
-        // test the value of the control against the regexp supplied
-        const valid = regex.test(control.value);
-    
-        // if true, return no error (no error), else return error passed in the second parameter
-        return valid ? null : error;
-    };
-}
 
 export function passwordMatchValidator(control: AbstractControl) {
     const password: string = control.get('password')!.value; // get password from our password form control
@@ -26,7 +9,3 @@ export function passwordMatchValidator(control: AbstractControl) {
         control.get('confirmPassword')!.setErrors({ NoPassswordMatch: true });
     }
 }
-
-
-
-
